+++ plugins/livereload.ts
@@ -5,11 +5,15 @@ export function livereload() {
     return function(this: Buchta | BuchtaCLI) {
         if (this instanceof Buchta) {
             this.livereload = {
-                clients: []
+                clients: [],
+                onUpdate: [],
+                registerOnUpdate: function(f: (file: string) => void) {
+                    this.onUpdate.push(f);
+                }
             }
 
             this.wsOnOpen(function(client) {
-                this.livereload.clients.push(client)
+                this.livereload.clients.push(client);
             })
         }
     }